+++ kernel/util/conv.c
@@ -21,8 +21,8 @@
 #include <tupai/util/conv.h>
 #include <tupai/util/math.h>
 
-static bool wrapper_signed(int64_t x, int base, int pad, int size, char* str);
-static bool wrapper_unsigned(uint64_t x, int base, int pad, int size, char* str);
+static bool wrapper_signed(intmax_t x, int base, int pad, int size, char* str);
+static bool wrapper_unsigned(uintmax_t x, int base, int pad, int size, char* str);
 
 bool i8_to_str(int8_t x, int base, int pad, char (str)[I8_STR_MAX])
 {
@@ -64,15 +64,15 @@ bool u64_to_str(uint64_t x, int base, int pad, char (str)[I64_STR_MAX])
 	return wrapper_unsigned(x, base, pad, sizeof(uint64_t), str);
 }
 
-bool ptr_to_str(size_t x, int base, int pad, char (str)[PTR_STR_MAX])
+bool ptr_to_str(uintptr_t x, int base, int pad, char (str)[PTR_STR_MAX])
 {
-	return wrapper_unsigned(x, base, pad, sizeof(size_t), str);
+	return wrapper_unsigned(x, base, pad, sizeof(uintptr_t), str);
 }
 
-bool wrapper_signed(int64_t x, int base, int pad, int size, char* str)
+bool wrapper_signed(intmax_t x, int base, int pad, int size, char* str)
 {
 	int dig = 0;
-	for (int32_t t = abs(x); t > 0; t /= base) dig ++;
+	for (intmax_t t = abs(x); t > 0; t /= base) dig ++;
 
 	if (x < 0)
 	{
@@ -106,10 +106,10 @@ bool wrapper_signed(int64_t x, int base, int pad, int size, char* str)
 	return true;
 }
 
-bool wrapper_unsigned(uint64_t x, int base, int pad, int size, char* str)
+bool wrapper_unsigned(uintmax_t x, int base, int pad, int size, char* str)
 {
 	int dig = 0;
-	for (size_t t = x; t > 0; t /= base) dig ++;
+	for (uintmax_t t = x; t > 0; t /= base) dig ++;
 
 	for (int i = min(size * 8, max(1, pad)) - dig; i > 0; i --)
 		*(str++) = '0';
